Delete README.md
You’re provided a MIPS assembly program, Assn4_MYMSUID.asm. In an exception to my usual practice, I’ve put the data segment below the text portion because the data is so long and I don’t want it to obscure your view of the program. That order of the two segments does not affect the behavior of the program. 

Each of the 5000 words in array1 in the provided program is the ASCII code for one of the characters in a story (see the comments at the end of each line). 5000 is a number in decimal notation. The data contains spaces (ASCII 0x20) and digits, but certainly no unprintable characters.

Your 3 tasks:

1.	Count the number of times each character appears in the provided data, regardless of whether the character is uppercase or lowercase.  

Do not use the equivalent of 26 if statements to increment that count.
# Don’t use this technique or some variation of it
if (character == A || character == a) then . . .
else if (character == B || character == b) then . . .
else if . . .

Instead, use arithmetic to determine the character’s appropriate index.
# Suppose the character actually is E, and that register $s7 
# contains the address of array results.
# Treating the characters as numbers, ‘E’ – ‘A’ = 5
# so your program should increment the count at
# the fifth array location
# NOTE: You must not use the address 20($s7) for that 5th 
# location, but instead 0($s6) where 
# $s6 = <address of array1> + (array_index * 4)

2.	Store the count for each character into an array results. For instance, supposing that the register $s7 contains the address of array results, then:
o	the word at 0($s7) should contain the count for the character A or a,
o	the word at 4($s7) should contain the count for the character B or b, 
o	etc.

3.	Print a message announcing the characters with the five highest number of occurrences. 
o	Print something like “The number of occurrences of a character is E with 42.” (That’s merely an example – neither E nor 42 may be the genuine answer.) 
o	Do not use the equivalent of 26 if statements to print the specific character. However, you may wish to use the equivalent of 26 data statements or initializations.